+++ app/main.py
@@ -1,10 +1,10 @@
 from fastapi import FastAPI
 
-from .routers import gs_mock, hello
+from .routers import gs, hello
 
 app = FastAPI()
 
-app.include_router(gs_mock.router, prefix="/api/v1")
+app.include_router(gs.router, prefix="/api/v1")
 app.include_router(hello.router, prefix="/api/v1")
